+++ src/store/index.ts
@@ -34,6 +34,8 @@ export {
   updateTaskInTree
 } from './TaskTreeOps'
 export type { FlatTask } from './TaskTreeOps'
+export { VaultIndex } from './VaultIndex'
+export type { ProjectRef, TaskRef } from './VaultIndex'
 export { projectPathForTaskPath } from './vaultFs'
 export { hydrateTasks } from './YamlHydrator'
 export { appendYaml, isOldFormat, parseFrontmatter } from './YamlParser'